+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.20.0-nullsafety.4
 
-- Adds a dry-run mode to check if an SVG is supported.
+- Adds option `warningsAsErrors` that throws errors when detecting unsupported SVG elements.
 
 ## 0.20.0-nullsafety.3